+++ designer/client/src/components/tips/Warnings.tsx
@@ -5,7 +5,7 @@ import NodeUtils from "../graph/NodeUtils";
 import { groupBy } from "lodash";
 import { LinkStyled } from "./Styled";
 import { NodeType, ScenarioGraph } from "../../types";
-import { styled } from "@mui/material";
+import { styled, Typography } from "@mui/material";
 
 interface Warning {
     error: {
@@ -43,17 +43,22 @@ const Warnings = ({ warnings, showDetails, scenarioGraph }: WarningsProps) => {
             <div>
                 {Object.entries(groupedByType).map(([warningType, warnings]) => (
                     <div key={uuid4()} title={warnings[0]?.error.description}>
-                        <span>{headerMessageByWarningType.get(warningType)}</span>
+                        <Typography component={"span"} variant={"body2"}>
+                            {headerMessageByWarningType.get(warningType)}
+                        </Typography>
                         <div style={{ display: "inline" }}>
                             {warnings.map((warning, index) => (
-                                <LinkStyled
+                                <Typography
+                                    variant={"body2"}
+                                    fontWeight={"bold"}
+                                    component={LinkStyled}
                                     key={uuid4()}
                                     to={""}
                                     onClick={(event) => showDetails(event, NodeUtils.getNodeById(warning.key, scenarioGraph))}
                                 >
-                                    <span>{warning.key}</span>
+                                    {warning.key}
                                     {index < warnings.length - 1 ? separator : null}
-                                </LinkStyled>
+                                </Typography>
                             ))}
                         </div>
                     </div>