+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/services/BaseHttpService.scala
@@ -1,12 +1,13 @@
 package pl.touk.nussknacker.ui.services
 
+import cats.data.EitherT
 import com.typesafe.config.Config
 import pl.touk.nussknacker.restmodel.SecurityError
-import SecurityError.{AuthenticationError, AuthorizationError}
+import pl.touk.nussknacker.restmodel.SecurityError.{AuthenticationError, AuthorizationError}
 import pl.touk.nussknacker.security.AuthCredentials
 import pl.touk.nussknacker.ui.security.api._
-import pl.touk.nussknacker.ui.services.BaseHttpService.NoRequirementServerEndpoint
-import sttp.tapir.server.ServerEndpoint
+import pl.touk.nussknacker.ui.services.BaseHttpService.{CustomAuthorizationError, NoRequirementServerEndpoint}
+import sttp.tapir.server.{PartialServerEndpoint, ServerEndpoint}
 
 import java.util.concurrent.atomic.AtomicReference
 import scala.concurrent.{ExecutionContext, Future}
@@ -71,10 +72,38 @@ abstract class BaseHttpService(
   protected def businessError[BUSINESS_ERROR](error: BUSINESS_ERROR) = Left(Left(error))
 
   protected def securityError[SE <: SecurityError](error: SE) = Left(Right(error))
+
+  private type PartialEndpoint[INPUT, OUTPUT, BUSINESS_ERROR, -R] =
+    PartialServerEndpoint[_, LoggedUser, INPUT, Either[BUSINESS_ERROR, SecurityError], OUTPUT, R, Future]
+
+  implicit class ServerLogicExtension[INPUT, OUTPUT, BUSINESS_ERROR, -R](
+      endpoint: PartialEndpoint[INPUT, OUTPUT, BUSINESS_ERROR, R]
+  ) {
+
+    def serverLogicEitherT(f: LoggedUser => INPUT => EitherT[Future, BUSINESS_ERROR, OUTPUT]) =
+      endpoint.serverLogic { loggedUser: LoggedUser => input: INPUT =>
+        f(loggedUser)(input).value.map(toTapirResponse)
+      }
+
+    private def toTapirResponse(result: Either[BUSINESS_ERROR, OUTPUT]): LogicResult[BUSINESS_ERROR, OUTPUT] =
+      result match {
+        case Left(error) =>
+          error match {
+            case _: CustomAuthorizationError => securityError(AuthorizationError)
+            case e                           => businessError(e)
+          }
+        case Right(value) => success(value)
+      }
+
+  }
+
 }
 
 object BaseHttpService {
 
   // we assume that our endpoints have no special requirements (in the Tapir sense)
   type NoRequirementServerEndpoint = ServerEndpoint[Any, Future]
+
+  // it's marker interface which simplifies error handling when serverLogicEitherT is used
+  trait CustomAuthorizationError
 }